+++ r-package/R/read_access.R
@@ -10,10 +10,14 @@
 #'             `city="all"`, results for all cities are loaded.
 #' @param mode Character. A transport mode. Modes available include
 #'             'public_transport', 'bicycle', or 'walk' (the default).
+#' @param peak Logical. If `TRUE` (the default), returns accessibility estimates
+#'             during peak time, between 6am and 8am. If `FALSE`, returns
+#'             accessibility during off-peak, between 2pm and 4am. This argument
+#'             only takes effect when `mode = public_transport`.
 #' @param year Numeric. A year number in YYYY format. Default set to 2019, the
 #'             only year currently available.
-#' @param geometry Logical. if FALSE (the default), returns a regular data.table
-#'                 of aop data. if TRUE, returns a an `sf data.frame` with simple
+#' @param geometry Logical. If `FALSE` (the default), returns a regular data.table
+#'                 of aop data. If `TRUE`, returns a an `sf data.frame` with simple
 #'                 feature geometry of spatial hexagonal grid H3. See details in
 #'                 \link{read_grid}.
 #' @param showProgress Logical. Defaults to `TRUE` display progress bar
@@ -31,7 +35,7 @@
 #' all <- read_access(city = 'all', mode = 'public_transport', year = 2019)
 
 #'}
-read_access <- function(city, mode = 'walk', year = 2019, geometry = FALSE, showProgress = TRUE){
+read_access <- function(city, mode = 'walk', peak = TRUE, year = 2019, geometry = FALSE, showProgress = TRUE){
 
   # Get metadata with data url addresses
   temp_meta <- select_metadata(t='access',
@@ -45,15 +49,23 @@ read_access <- function(city, mode = 'walk', year = 2019, geometry = FALSE, show
   # download files
   aop_df <- download_data(file_url, progress_bar = showProgress)
 
-  # without spatial data, return df
-  if(geometry == FALSE){
-                        return(aop_df)
-                        }
+  # peak Vs off-peak
+  if(peak==TRUE){
+                 aop_df <- subset(aop_df, peak == 1)
+                } else {
+                 aop_df <- subset(aop_df, peak == 0)
+                }
 
-  # with spatial data, return sf
-  if(geometry == TRUE){
+  # with Vs without spatial data
+  if(geometry == FALSE){
+                        # return df
+                        return(aop_df)
+
+                        } else {
+
+                        # return sf
                         aop_sf <- read_grid(city=city, showProgress=showProgress)
                         aop <- aop_spatial_join(aop_df, aop_sf)
                         return(aop)
                         }
-}
+  }