+++ r-package/NEWS.md
@@ -4,8 +4,12 @@
 
 # aopdata v0.1.0 (dev)
 
+
+**Major changes**
+* New parameter `peak` added to `read_access()` function. Closes #17.
+
 **Minor changes**
-* Downloads two or more cities at the same time. Closes #3
+* Downloads two or more cities at the same time. Closes #3.